cookie-consent onetrust
- added helper function to load scripts to head section of the page
- delayed cookie-consent onetrust loading via placeholders
- for /pages/zemax-cookie-policy, added an click event listener to attach the cookie options
