+++ scripts/delayed.js
@@ -1,7 +1,33 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM } from './lib-franklin.js';
+import { loadScript } from './scripts.js';
+import { sampleRUM, fetchPlaceholders } from './lib-franklin.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
 
-// add more delayed functionality here
+// fetch placeholders file
+const placeholders = await fetchPlaceholders();
+
+// OneTrust Cookies Consent Notice start
+const otId = placeholders.onetrustid;
+if (otId) {
+  loadScript('https://cdn.cookielaw.org/scripttemplates/otSDKStub.js', {
+    type: 'text/javascript',
+    charset: 'UTF-8',
+    'data-domain-script': `${otId}`,
+  });
+
+  window.OptanonWrapper = () => {};
+}
+
+const allButtons = document.querySelectorAll('a.button');
+allButtons.forEach((button) => {
+  if (button.getAttribute('href').includes('cookie-policy')) {
+    button.addEventListener('click', (e) => {
+    // eslint-disable-next-line no-undef
+      OneTrust.ToggleInfoDisplay();
+      e.preventDefault();
+    });
+  }
+});
+// OneTrust Cookies Consent Notice end
